fix(evals): isolate Harbor LangGraph project
