--- libs/evals/deepagents_harbor/langgraph.json
@@ -1,6 +1,0 @@
-{
-  "dependencies": ["."],
-  "graphs": {
-    "deepagent": "./deepagents_harbor/langgraph_agent.py:make_graph"
-  }
-}
