+++ libs/evals/deepagents_harbor/langgraph_project/langgraph.json
@@ -1,0 +1,14 @@
+{
+  "dependencies": [
+    "deepagents>=0.6.8",
+    "deepagents-code>=0.1.11",
+    "langchain>=1.3.4,<2.0.0",
+    "langchain-anthropic>=1.4.4,<1.5.0",
+    "aiohttp>=3.14.0,<4.0.0",
+    "toml>=0.10.2,<1.0.0"
+  ],
+  "graphs": {
+    "deepagent": "./langgraph_agent.py:make_graph"
+  }
+}
+
